requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
email-validator
